@@ -1,9 +1,10 @@
 name: example
 description: Example app using routing_path package
 version: 1.0.0+1
+publish_to: none
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
