@@ -4,16 +4,17 @@ version: 0.0.1+3
 homepage: https://github.com/williamhjcho/routing_path
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
+  build_runner: ^1.11.0
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.0
-  matcher: ^0.12.0
+  matcher: ^0.12.10-nullsafety.1
+  mockito: ^5.0.0
 
 flutter:
