@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:routing_path/routing_path.dart';
 
 class FeatureAHomePage extends StatelessWidget {
-  const FeatureAHomePage({Key key, @required this.openedBy}) : super(key: key);
+  const FeatureAHomePage({Key? key, required this.openedBy}) : super(key: key);
 
   final String openedBy;
 
@@ -21,9 +21,11 @@ class FeatureAHomePage extends StatelessWidget {
           children: <Widget>[
             Text('This page was opened by "$openedBy"'),
             const SizedBox(height: 16),
-            OutlineButton(
-              textColor: Colors.indigo,
-              onPressed: () => Router.of(context).open(
+            OutlinedButton(
+              style: ButtonStyle(
+                foregroundColor: MaterialStateProperty.all(Colors.indigo),
+              ),
+              onPressed: () => PathRouter.of(context).open(
                 '/feature-b',
                 RouteArguments({'opened_by': 'FEATURE A'}),
               ),
