@@ -4,7 +4,7 @@ import 'route_handler.dart';
 
 /// The base interface for opening [RouteHandler]s.
 ///
-/// Usually there should be one [Router] instance declared close to the root
+/// Usually there should be one [PathRouter] instance declared close to the root
 /// widget to allow others to call any arbitrary routes from a [BuildContext].
 ///
 /// Typical usage is as follows:
@@ -12,8 +12,8 @@ import 'route_handler.dart';
 /// ```dart
 /// Router.of(context).open(...);
 /// ```
-abstract class Router extends StatelessWidget implements RouteHandler {
-  const Router({Key key, this.child}) : super(key: key);
+abstract class PathRouter extends StatelessWidget implements RouteHandler {
+  const PathRouter({Key? key, required this.child}) : super(key: key);
 
   /// {@macro flutter.widgets.child}
   final Widget child;
@@ -26,18 +26,38 @@ abstract class Router extends StatelessWidget implements RouteHandler {
   /// encloses the given [context].
   ///
   /// May return null if not found.
-  static Router of(BuildContext context) =>
-      context.dependOnInheritedWidgetOfExactType<_InheritedRouter>()?.router;
+  static PathRouter of(BuildContext context) {
+    final inheritedRouter =
+        context.dependOnInheritedWidgetOfExactType<_InheritedRouter>();
+    assert(() {
+      if (inheritedRouter == null) {
+        throw FlutterError(
+          'Router operation requested with a context that does not include '
+          'a Router.\n'
+          'The context used to access the Router must be that of a widget that '
+          'is a descendant of a Router widget.',
+        );
+      }
+      return true;
+    }());
+    return inheritedRouter!.router;
+  }
+
+  static PathRouter? maybeOf(BuildContext context) {
+    final inheritedRouter =
+        context.dependOnInheritedWidgetOfExactType<_InheritedRouter>();
+    return inheritedRouter?.router;
+  }
 }
 
 class _InheritedRouter extends InheritedWidget {
   const _InheritedRouter({
-    Key key,
-    @required this.router,
-    Widget child,
+    Key? key,
+    required this.router,
+    required Widget child,
   }) : super(key: key, child: child);
 
-  final Router router;
+  final PathRouter router;
 
   @override
   bool updateShouldNotify(_InheritedRouter oldWidget) =>
