@@ -2,19 +2,20 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:routing_path/routing_path.dart';
 
-class ConcreteRouter extends Router {
-  const ConcreteRouter({Key key, Widget child}) : super(key: key, child: child);
+class ConcreteRouter extends PathRouter {
+  const ConcreteRouter({Key? key, required Widget child})
+      : super(key: key, child: child);
 
   @override
   bool canOpen(String path) => false;
 
   @override
-  Future<T> open<T>(String path, [dynamic arguments]) =>
+  Future<T?> open<T>(String path, [RouteArguments? arguments]) =>
       Future.error('Some error');
 }
 
 void main() {
-  BuildContext capturedContext;
+  BuildContext? capturedContext;
 
   final builder = Builder(builder: (context) {
     capturedContext = context;
@@ -34,17 +35,17 @@ void main() {
     testWidgets('when in hierarchy', (tester) async {
       final router = ConcreteRouter(child: builder);
       await tester.pumpWidget(router);
-      expect(Router.of(capturedContext), same(router));
+      expect(PathRouter.of(capturedContext!), same(router));
     });
 
     testWidgets('when not in hierarchy', (tester) async {
       await tester.pumpWidget(builder);
-      expect(Router.of(capturedContext), isNull);
+      expect(() => PathRouter.of(capturedContext!), throwsFlutterError);
     });
 
     testWidgets('when there are multiple routers in the hierarchy',
         (tester) async {
-      BuildContext rootContext, subContext;
+      late BuildContext rootContext, subContext;
 
       final subRouter = ConcreteRouter(child: Builder(builder: (context) {
         subContext = context;
@@ -56,8 +57,39 @@ void main() {
       }));
       await tester.pumpWidget(rootRouter);
 
-      expect(Router.of(rootContext), same(rootRouter));
-      expect(Router.of(subContext), same(subRouter));
+      expect(PathRouter.of(rootContext), same(rootRouter));
+      expect(PathRouter.of(subContext), same(subRouter));
+    });
+  });
+
+  group('.maybeOf', () {
+    testWidgets('when in hierarchy', (tester) async {
+      final router = ConcreteRouter(child: builder);
+      await tester.pumpWidget(router);
+      expect(PathRouter.maybeOf(capturedContext!), same(router));
+    });
+
+    testWidgets('when not in hierarchy', (tester) async {
+      await tester.pumpWidget(builder);
+      expect(PathRouter.maybeOf(capturedContext!), isNull);
+    });
+
+    testWidgets('when there are multiple routers in the hierarchy',
+        (tester) async {
+      late BuildContext rootContext, subContext;
+
+      final subRouter = ConcreteRouter(child: Builder(builder: (context) {
+        subContext = context;
+        return Container();
+      }));
+      final rootRouter = ConcreteRouter(child: Builder(builder: (context) {
+        rootContext = context;
+        return subRouter;
+      }));
+      await tester.pumpWidget(rootRouter);
+
+      expect(PathRouter.maybeOf(rootContext), same(rootRouter));
+      expect(PathRouter.maybeOf(subContext), same(subRouter));
     });
   });
 }
