@@ -2,39 +2,39 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:routing_path/src/utils/path_matcher.dart';
 
 void main() {
-  Map<String, String> match(String pathPattern, String path) {
+  Map<String, String>? matches(String pathPattern, String path) {
     final pattern = buildPathPattern(pathPattern);
     return pathMatches(pattern, path);
   }
 
   test('simple valid path matches', () {
     // leading and trailing slash should be consumed when matching
-    expect(match('', ''), {});
-    expect(match('/', '/'), {});
-    expect(match('/', ''), {});
+    expect(matches('', ''), {});
+    expect(matches('/', '/'), {});
+    expect(matches('/', ''), {});
 
-    expect(match('/path/to/somewhere', '/path/to/somewhere'), {});
-    expect(match('/path/to/somewhere', '/path/to/somewhere/'), {});
-    expect(match('/path/to/somewhere', 'path/to/somewhere'), {});
+    expect(matches('/path/to/somewhere', '/path/to/somewhere'), {});
+    expect(matches('/path/to/somewhere', '/path/to/somewhere/'), {});
+    expect(matches('/path/to/somewhere', 'path/to/somewhere'), {});
 
-    expect(match('/path_to-somewhere', '/path_to-somewhere'), {});
+    expect(matches('/path_to-somewhere', '/path_to-somewhere'), {});
   });
 
   test('simple invalid path matches', () {
-    expect(match('/path/to/somewhere', '/'), isNull);
-    expect(match('/path/to/somewhere', '/path'), isNull);
-    expect(match('/path/to/somewhere', '/path/to'), isNull);
+    expect(matches('/path/to/somewhere', '/'), isNull);
+    expect(matches('/path/to/somewhere', '/path'), isNull);
+    expect(matches('/path/to/somewhere', '/path/to'), isNull);
 
     // must be a exact match
-    expect(match('/path/to/somewhere', '/path/to/somewhereeee'), isNull);
-    expect(match('/path/to/somewhere', '/path/to/somewhere/here'), isNull);
+    expect(matches('/path/to/somewhere', '/path/to/somewhereeee'), isNull);
+    expect(matches('/path/to/somewhere', '/path/to/somewhere/here'), isNull);
   });
 
   test('replaceable valid path matches', () {
-    expect(match('/:id', '/123'), {'id': '123'});
+    expect(matches('/:id', '/123'), {'id': '123'});
 
     expect(
-      match(
+      matches(
         '/:id/:name/:value',
         '/123/path-name/456-789',
       ),
@@ -46,7 +46,7 @@ void main() {
     );
 
     expect(
-      match(
+      matches(
         '/path/:id/:name',
         '/path/123/path-name',
       ),
@@ -56,7 +56,7 @@ void main() {
       },
     );
     expect(
-      match(
+      matches(
         '/:id/path/:name',
         '/123/path/path-name',
       ),
@@ -68,21 +68,21 @@ void main() {
   });
 
   test('replaceable invalid path matches', () {
-    expect(match('/', '/:id'), isNull);
-    expect(match('/123', '/:id'), isNull);
+    expect(matches('/', '/:id'), isNull);
+    expect(matches('/123', '/:id'), isNull);
 
-    expect(match('/:id', '/'), isNull);
-    expect(match('/:id', '/123/456'), isNull);
+    expect(matches('/:id', '/'), isNull);
+    expect(matches('/:id', '/123/456'), isNull);
 
     expect(
-      match(
+      matches(
         '/path/:name/somewhere',
         '/path/path-name',
       ),
       isNull,
     );
     expect(
-      match(
+      matches(
         '/path/:id/to/:name',
         '/path/123/path-name',
       ),
